language_api/requirements.txt
Flask==1.1.2
json2html==1.3.0
pandas==1.1.5
numpy==1.19.5
matplotlib==3.3.4
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-cloud-datastore==2.1.0
google-cloud-language==2.0.0